RAG/requirements.txt
openai
faiss-cpu
PyPDF2
streamlit
numpy
